requirements.txt
# Core
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Optional for async (we'll use simple version first)
# aiohttp==3.9.1

# Testing
pytest==7.4.3

# Utils
tqdm==4.66.1  # Progress bars